Renamed pictures and Added code
just renamed edited images ans 400px and 120px as px as pixel to be more specific and added code to display image of fire fox and using src which means source of fine and alt to provide  text for image which is going to be displayed if the image fails to load and srcset define set of possible image source and their respective widths for responsive image and i have mention
firefox_logo-only_RGB_400px.png px for 400 pixel and firefox_logo-only_RGB_120px.png" for 120 pixel and using sizes to size different view point for maximum width of 500 pixel the image should be displayed as 120 pixels; otherwise, it should be 400 pixels. This helps in creating a responsive design for different screen sizes.