Update blank-template.html
Added data for Ice giants (Uranus and Neptune) and Dwarf planets (Pluto) to the HTML table. This includes information on mass, diameter, density, gravity, length of day, distance from the Sun, mean temperature, number of moons, and notes. The table provides a comprehensive overview of various planetary categories in our solar system.
Added information about Pluto's declassification as a planet in 2006, with a hyperlink to a USA Today article discussing the controversial nature of this decision.